+++ inst/shinyApp/ui.R
@@ -143,7 +143,6 @@ shinyUI(fluidPage(
     tags$script(src = "svg-parser/svg-parser-bundle.js"),
     tags$script(src = "scale-that-svg/scale-that-svg-bundle.js"),
     tags$script(src = "word-wrap/word-wrap-bundle.js"),
-#    tags$script(src = "xml-formatter/xml-formatter_bundle.js"),
     tags$script(src = "panzoom/panzoom.min.js"),
     tags$link(rel = "stylesheet", href = "chrome-tabs/chrome-tabs.css"),
     tags$link(rel = "stylesheet", href = "chrome-tabs/chrome-tabs-dark-theme.css"),