+++ NEWS.md
@@ -1,4 +1,4 @@
-# shinyMonacoEditor 1.0.0.9000
+# shinyMonacoEditor 1.1.0
 
 * Haskell formatter, requires `brittany`.
 * XML formatter requiring `xmllint`.
@@ -9,7 +9,7 @@
 
 # shinyMonacoEditor 1.0.0
 
-* Updated 'prettier'.
+* Updated `prettier`.
 * A trailing comma was added in JavaScript objects when prettifying JavaScript or HTML; this does not happen anymore.
 * New feature: SVG rescaling. 
 * New feature: SVG panning and zooming.